+++ src/main/java/me/pugabyte/nexus/features/minigames/models/perks/common/ParticleProjectilePerk.java
@@ -5,8 +5,8 @@ import me.pugabyte.nexus.features.minigames.models.Match;
 import me.pugabyte.nexus.features.minigames.models.Minigamer;
 import me.pugabyte.nexus.features.minigames.models.perks.Perk;
 import me.pugabyte.nexus.features.minigames.models.perks.PerkCategory;
-import me.pugabyte.nexus.features.minigames.models.perks.PerkOwner;
-import me.pugabyte.nexus.features.minigames.models.perks.PerkOwnerService;
+import me.pugabyte.nexus.models.perkowner.PerkOwner;
+import me.pugabyte.nexus.models.perkowner.PerkOwnerService;
 import org.bukkit.Particle;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Projectile;
